fixing bug in Register()
when RegisterWithDB was created, the setting of the driver variable was moved from Register, this caused the opening of the SQL connection to not have a driver string ultimately causing this call to fail. since driver is created outside of the function scope, this passes a compile. 

replaced 'driver' with 'driverName' which fixes the issue